fix a VC7 warning in ogg file seeking stuff
git-svn-id: http://audiere.svn.sourceforge.net/svnroot/audiere/trunk@129 cb5fa4c3-b70d-0410-b4f6-90f5898e5689

+++ input_ogg.cpp
@@ -173,7 +173,7 @@ OGGInputStream::FileSeek(void* opaque, int64_t offset, int whence)
     case SEEK_END: type = ADR_END;     break;
     default: return -1;
   }
-  return (file->Seek(offset, type) ? 0 : -1);
+  return (file->Seek((int)offset, type) ? 0 : -1);
 }
 
 ////////////////////////////////////////////////////////////////////////////////
